The student has become the master
